+++ packages/commercetools/api-client/__tests__/api/setShippingMethod.spec.ts
@@ -1,4 +1,4 @@
-import { setShippingMethodAction } from './../../../src/helpers/cart/actions';
+import { setShippingMethodAction } from '../../src/helpers/cart/actions';
 
 describe('[commercetools-api-client] setShippingMethod', () => {
   beforeEach(() => {
