+++ packages/commercetools/api-client/__tests__/api/updateShippingDetails.spec.ts
@@ -1,4 +1,4 @@
-import updateShippingDetails from './../../../src/api/updateShippingDetails';
+import updateShippingDetails from '../../src/api/updateShippingDetails';
 
 const cart = {
   id: 1,
