+++ packages/commercetools/api-client/__tests__/setup.ts
@@ -1,11 +1,11 @@
 import { createApiClient } from './../src/index';
 
-jest.mock('./../src/helpers/createCommerceToolsLink');
-jest.mock('./../src/api/updateCart', () => jest.fn((arg) => arg));
-jest.mock('./../src/api/createMyOrderFromCart', () => jest.fn((arg) => arg));
+jest.mock('../src/helpers/createCommerceToolsLink');
+jest.mock('../src/api/updateCart', () => jest.fn((arg) => arg));
+jest.mock('../src/api/createMyOrderFromCart', () => jest.fn((arg) => arg));
 jest.mock('apollo-client');
 jest.mock('@commercetools/sdk-auth');
-jest.mock('./../src/helpers/createAccessToken', () => jest.fn());
+jest.mock('../src/helpers/createAccessToken', () => jest.fn());
 
 export default createApiClient({
   api: {} as any,
