+++ packages/commercetools/api-client/src/helpers/apiClient/index.ts
@@ -1,0 +1,74 @@
+import ApolloClient from 'apollo-client';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { apiClientFactory } from '@vue-storefront/core';
+import * as api from '../../api';
+import { Config, ClientInstance } from '../../types/setup';
+import { createCommerceToolsConnection } from '../../helpers/commercetoolsLink';
+
+const defaultSettings = {
+  locale: 'en',
+  acceptLanguage: ['en'],
+  auth: {
+    onTokenChange: () => {},
+    onTokenRead: () => '',
+    onTokenRemove: () => {}
+  },
+  cookies: {
+    currencyCookieName: 'vsf-currency',
+    countryCookieName: 'vsf-country',
+    localeCookieName: 'vsf-locale'
+  }
+};
+
+const onSetup = (settings: Config): { config: Config; client: ClientInstance } => {
+  const languageMap = settings.languageMap || {};
+  const acceptLanguage = settings.acceptLanguage || defaultSettings.acceptLanguage;
+  const locale = settings.locale || defaultSettings.locale;
+
+  const config = {
+    ...defaultSettings,
+    ...settings,
+    languageMap,
+    acceptLanguage: languageMap[locale] || acceptLanguage,
+    auth: settings.auth || defaultSettings.auth
+  } as any as Config;
+
+  if (settings.client) {
+    return { client: settings.client, config };
+  }
+
+  if (settings.customOptions && settings.customOptions.link) {
+    return {
+      client: new ApolloClient({
+        cache: new InMemoryCache(),
+        ...settings.customOptions
+      }),
+      config
+    };
+  }
+
+  const { apolloLink, sdkAuth, tokenProvider } = createCommerceToolsConnection(config);
+
+  const client = new ApolloClient({
+    link: apolloLink,
+    cache: new InMemoryCache(),
+    ...settings.customOptions
+  });
+  (client as ClientInstance).sdkAuth = sdkAuth;
+  (client as ClientInstance).tokenProvider = tokenProvider;
+
+  return {
+    config,
+    client
+  };
+};
+
+const { createApiClient } = apiClientFactory<Config, any>({
+  tag: 'ct',
+  onSetup,
+  api
+});
+
+export {
+  createApiClient
+};
