+++ packages/commercetools/api-client/src/helpers/commercetoolsLink/helpers.ts
