+++ packages/commercetools/api-client/src/types/Api.ts
@@ -1,7 +1,52 @@
 import { ApolloQueryResult } from 'apollo-client';
 import { FetchResult } from 'apollo-link';
-import { Cart, Order, ShippingMethod, CustomerSignInResult, Customer, CartDraft } from './GraphQL';
 import { Token, CustomerCredentials } from './setup';
+import { UpdateCartParams } from '../api/updateCart';
+import { GetMeParams } from '../api/getMe';
+import { ShippingMethodData } from '../api/getShippingMethods';
+import {
+  Cart,
+  Order,
+  ShippingMethod,
+  CustomerSignInResult,
+  Customer,
+  CartDraft,
+  ProductVariant,
+  OrderMyCartCommand,
+  CustomerSignMeInDraft,
+  CustomerSignMeUpDraft,
+  ReferenceInput,
+  Address,
+  LineItem,
+  CategoryQueryResult,
+  ProductQueryResult,
+  Me,
+  CartQueryInterface
+} from './GraphQL';
+
+export interface ApiInstance {
+  addToCart ({ id, version }: Cart, product: ProductVariant, quantity: number, customQuery?: CustomQueryFn): Promise<CartResponse>;
+  applyCartCoupon (cart: Cart, discountCode: string, customQuery?: CustomQueryFn): Promise<CartResponse>;
+  createCart (cartDraft?: CartData, customQueryFn?: CustomQueryFn): Promise<{ data: CartQueryInterface }>;
+  createMyOrderFromCart (draft: OrderMyCartCommand, customQueryFn?: CustomQueryFn): Promise<OrderMutationResponse>;
+  customerChangeMyPassword (version: any, currentPassword: string, newPassword: string): Promise<ChangeMyPasswordResponse>;
+  customerSignMeIn (draft: CustomerSignMeInDraft): Promise<SignInResponse>;
+  customerSignMeUp (draft: CustomerSignMeUpDraft): Promise<SignInResponse>;
+  customerSignOut (): Promise<void>;
+  customerUpdateMe (currentUser, updatedUserData): Promise<any>;
+  getCart (cartId: string): Promise<CartQueryResponse>;
+  getCategory (params, customQueryFn?: CustomQueryFn): Promise<QueryResponse<'categories', CategoryQueryResult>>;
+  getMe (params?: GetMeParams, customQueryFn?: CustomQueryFn): Promise<{ data: { me: Me } }>;
+  getOrders (params, customQueryFn?: CustomQueryFn): Promise<{ data: { me: Me } }>;
+  getProduct (params, customQueryFn?: CustomQueryFn): Promise<QueryResponse<'products', ProductQueryResult>>;
+  getShippingMethods (cartId?: string, customQueryFn?: CustomQueryFn): Promise<ShippingMethodData>;
+  removeCartCoupon (cart: Cart, discountCode: ReferenceInput, customQuery?: CustomQueryFn): Promise<CartResponse>;
+  removeFromCart (cart: Cart, product: LineItem, customQuery?: CustomQueryFn): Promise<CartResponse>;
+  updateCart (params: UpdateCartParams, customQueryFn?: CustomQueryFn): Promise<CartResponse>;
+  updateCartQuantity (cart: Cart, product: LineItem, customQuery?: CustomQueryFn): Promise<CartResponse>;
+  updateShippingDetails (cart: Cart, shippingDetails: Address, customQueryFn?: CustomQueryFn): Promise<CartResponse>;
+  isGuest: () => boolean;
+}
 
 export type CustomQueryFn<T = any> = (query: any, variables: T) => {
   query?: any;
